n8n-utils/fastapi-proxy/pyproject.toml
[project]
name = "fastapi-proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "python-jobspy>=1.1.82",
]
